pyproject.toml
[tool.poetry]
name = "django-nextjs-backend-api"
version = "0.1.0"
description = ""
authors = ["nikhil <nikhil.bhardwaj@sveltetech.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
django-ninja = "^1.2.1"
rav = "^0.0.9"
django-ninja-jwt = {extras = ["crypto"], version = "^5.3.2"}
django-extensions = "^3.2.3"
ipython = "^8.26.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
